% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cnomal.R
\name{Cnormal}
\alias{Cnormal}
\title{Contaminated Normal Distribution}
\usage{
Cnormal(parmt = stop("no epsi or sigma argument"))
}
\arguments{
\item{parmt}{parameters vector (epsi, sigma).}
}
\value{
An object of class "family" for contaminated normal distribution.
}
\description{
Family objects for contaminated normal distribution provide a convenient way to specify the details of the models used by functions such as \code{elliptical}.
}
\examples{
\dontrun{
data(luzdat)
y <- luzdat$y
x1 <- luzdat$x1 ; x1 <- factor(x1) ; x1 <- C(x1,treatment)
x2 <- luzdat$x2
x3 <- (luzdat$x2)^2
luz <- data.frame(y,x1,x2,x3)
elliptical.fitCn <- elliptical(y ~ x1+x2+x3, family = Cnormal(c(1,.5))
,data=luz)
family(elliptical.fitCn)
}
}
\references{
Fang, K. T., Kotz, S. and NG, K. W. (1990, ISBN:9781315897943).
Symmetric Multivariate and Related Distributions. London: Chapman and Hall.
}
\seealso{
\code{\link{family.elliptical}}, \code{\link{elliptical}}, \code{\link{gwer}}
}
\keyword{distributions}
\keyword{elliptical}
