% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Student.R
\name{Student}
\alias{Student}
\title{t-Student Distribution}
\usage{
Student(df = stop("no df argument"))
}
\arguments{
\item{df}{degrees of freedom.}
}
\value{
An object of class "family" for student distribution.
}
\description{
Family objects for t-Student distribution provide a convenient way to specify the details of the models used by functions such as \code{elliptical}.
}
\examples{
data(luzdat)
y <- luzdat$y
x1 <- luzdat$x1 ; x1 <- factor(x1) ; x1 <- C(x1,treatment)
x2 <- luzdat$x2
x3 <- (luzdat$x2)^2
luz <- data.frame(y,x1,x2,x3)
elliptical.fitt <- elliptical(y ~ x1+x2+x3, family = Student(df=5)
,data=luz)
family(elliptical.fitt)
}
\references{
Fang, K. T., Kotz, S. and NG, K. W. (1990, ISBN:9781315897943).
Symmetric Multivariate and Related Distributions. London: Chapman and Hall.
}
\seealso{
\code{\link{family.elliptical}}, \code{\link{elliptical}}, \code{\link{gwer}}
}
\keyword{distributions}
\keyword{elliptical}
